pyproject.toml
[tool.poetry]
name = "aimeer"
version = "0.1.0"
description = ""
authors = ["Hikki02 <a.aidaraliev25@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.2.1"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.6"
Pillow = "^9.5.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.0.0"
drf-yasg = "^1.21.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
